extract getAllSales
