+++ monolith/routes/auth.ts
@@ -3,9 +3,9 @@ const router = express.Router();
 import { MongoClient, ObjectID } from 'mongodb';
 import jwt from 'jsonwebtoken';
 import fetchDbName from '../lib/fetchDbName';
+const DATABASE_NAME = fetchDbName();
 import getCardsByFilter, { Arguments } from '../interactors/getCardsByFilter';
 import addCardToInventory from '../interactors/addCardToInventory';
-const DATABASE_NAME = fetchDbName();
 require('dotenv').config();
 import { Request } from 'express';
 import { ClubhouseLocation } from '../interactors/getJwt';
@@ -14,6 +14,7 @@ import collectionFromLocation from '../lib/collectionFromLocation';
 import getCardsWithInfo from '../interactors/getCardsWithInfo';
 import finishSale from '../interactors/updateInventoryCards';
 import getSalesFromCardname from '../interactors/getSalesFromCardname';
+import getAllSales from '../interactors/getAllSales';
 
 interface RequestWithUserInfo extends Request {
     locations: string[];
@@ -162,56 +163,6 @@ router.post('/finishSale', async (req: RequestWithUserInfo, res) => {
         res.status(500).send(err.message);
     }
 });
-
-async function getAllSales(location: ClubhouseLocation) {
-    const client = await new MongoClient(process.env.MONGO_URI, mongoOptions);
-
-    try {
-        await client.connect();
-        const db = client.db(DATABASE_NAME);
-
-        const data = await db
-            .collection(collectionFromLocation(location).salesData)
-            .find({})
-            .project({
-                sale_data: 1,
-                'sale_data.total': 1,
-                'sale_data.saleID': 1,
-                'sale_data.timeStamp': 1,
-                card_list: 1,
-                'card_list.foil': 1,
-                'card_list.nonfoil': 1,
-                'card_list.id': 1,
-                'card_list.name': 1,
-                'card_list.set': 1,
-                'card_list.set_name': 1,
-                'card_list.rarity': 1,
-                'card_list.price': 1,
-                'card_list.finishCondition': 1,
-                'card_list.reserved': 1,
-                'card_list.qtyToSell': 1,
-                'card_list.card_faces': 1,
-            });
-
-        const docs = await data.toArray();
-
-        // Some data, like 'total' and 'price' should be coerced to a Number for ease of use on the frontend
-        docs.forEach((el) => {
-            el.sale_data.total = Number(el.sale_data.total);
-
-            el.card_list.forEach((el, idx, arr) => {
-                arr[idx].price = Number(arr[idx].price);
-            });
-        });
-
-        return docs;
-    } catch (err) {
-        console.log(err);
-        throw err;
-    } finally {
-        await client.close();
-    }
-}
 
 async function getFormatLegalities(location: ClubhouseLocation) {
     const client = await new MongoClient(process.env.MONGO_URI, mongoOptions);
